+++ lib/types.ts
@@ -40,7 +40,6 @@ export interface IResult<T, D = string, M = {}> {
  */
 export type Payload<T, D = string, M = {}> = IResult<T, D, M>;
 export type HandlerPayload<T> = { aggregate: T, eventName: string };
-export type EventHandler<T, B> = ICommand<HandlerPayload<T>, Promise<B> | B>;
 
 /**
  * 
@@ -179,51 +178,6 @@ export interface IPublicHistory<Props> {
 
 export type IPropsValidation<T> = { [P in keyof Required<T>]: (value: T[P]) => boolean };
 
-/**
- * @description Domain Events Params
- * @param aggregate the entity to add the event.
- * @param createdAt the current date time the event was created.
- * @param callback the event handler to be executed on dispatch.
- */
-export interface IDomainEvent<T> {
-	aggregate: T;
-	createdAt: Date;
-	callback: IHandle<T>;
-}
-
-/**
- * @description Define a handler to be executed on dispatch an event.
- * @var eventName is optional value as string or undefine.
- * @method dispatch is the method to be executed on dispatch the event.
- */
-export interface IHandle<T> {
-	/**
-	 * @description eventName is optional value. Default is className
-	 */
-	eventName?: string;
-	dispatch(event: IDomainEvent<T>, handler: EventHandler<T, void>): Promise<void> | void;
-}
-
-/**
- * @description Options to dispatch some event.
- * @param eventName is the value defined on handler. Default is the class name.
- * @param id is the ID to identify the aggregate on state.
- */
-export interface IDispatchOptions {
-	eventName: string;
-	id: UID<string>;
-}
-
-/**
- * @description Event options
- */
-export interface IEvent<G> {
-	event: IDomainEvent<G>;
-	replace?: boolean;
-}
-
-export type IReplaceOptions = 'REPLACE_DUPLICATED' | 'UPDATE' | 'KEEP';
-
 export interface IAdapter<F, T, E = any, M = any> {
 	build(target: F): IResult<T, E, M>;
 }
@@ -256,7 +210,6 @@ export interface IAggregate<Props> {
 	hashCode(): UID<string>;
 	isNew(): boolean;
 	clone(): IEntity<Props>;
-	addEvent(event: IHandle<IAggregate<Props>>, replace?: IReplaceOptions): void;
 	deleteEvent(eventName: string): void;
 }
 
@@ -297,4 +250,104 @@ export interface EventMetrics {
 	current: number;
 	total: number;
 	dispatch: number;
+}
+
+/**
+ * Interface representing an event.
+ */
+export interface DEvent<T> {
+	eventName: string;
+	handler: Handler<T>;
+	options: Options;
+}
+
+export type HandlerArgs<T> = [T, [DEvent<T>, ...any[]]]
+
+/**
+ * Represents a promise-based event handler.
+ */
+export type PromiseHandler<T> = (...args: HandlerArgs<T>) => Promise<void>;
+
+/**
+ * Represents a normal event handler.
+ */
+export type NormalHandler<T> = (...args: HandlerArgs<T>) => void;
+
+/**
+ * Represents an event handler, which can be either a promise-based handler or a normal handler.
+ */
+export type Handler<T> = PromiseHandler<T> | NormalHandler<T>;
+
+/**
+ * Interface representing options for an event.
+ */
+export interface Options {
+	priority: number;
+}
+
+/**
+ * Interface representing metrics related to events.
+ * @interface Metrics
+ */
+export interface Metrics {
+	/**
+	 * A function that returns the total number of events.
+	 * @returns {number} The total number of events.
+	 */
+	totalEvents: () => number;
+
+	/**
+	 * A function that returns the total number of dispatched events.
+	 * @returns {number} The total number of dispatched events.
+	 */
+	totalDispatched: () => number;
+}
+
+/**
+ * Parameters for defining an event.
+ */
+export interface EventParams {
+    /**
+     * The name of the event.
+     */
+    eventName: string;
+    /**
+     * Additional options for the event.
+     */
+    options?: Options;
+}
+
+
+/**
+ * Abstract class representing an event handler.
+ * @template T - The type of aggregate this event handler is associated with.
+ */
+export abstract class EventHandler<T> {
+    /**
+     * Creates an instance of EventHandler.
+     * @param {EventParams} params - Parameters for the event handler.
+     * @throws {Error} If params.eventName is not provided as a string.
+     */
+    constructor(public readonly params: EventParams) {
+        if (typeof params?.eventName !== 'string') {
+            throw new Error('params.eventName is required as string');
+        }
+    }
+
+    /**
+     * Dispatches the event with the provided arguments.
+     * @abstract
+     * @param {T} aggregate - The aggregate associated with the event.
+     * @param {[DEvent<T>, any[]]} args - Arguments for the event dispatch.
+     * @returns {Promise<void> | void} A Promise if the event dispatch is asynchronous, otherwise void.
+     */
+    abstract dispatch(aggregate: T, args: [DEvent<T>, any[]]): Promise<void> | void;
+
+    /**
+     * Dispatches the event with the provided arguments.
+     * @abstract
+     * @param {...HandlerArgs<T>} args - Arguments for the event dispatch.
+     * @returns {Promise<void> | void} A Promise if the event dispatch is asynchronous, otherwise void.
+     */
+    abstract dispatch(...args: HandlerArgs<T>): Promise<void> | void;
 }